Cleanup tests running many times
