+++ test/sanityQueryService.test.js
@@ -1,10 +1,7 @@
-import {initializeService} from '../src/services/config.js';
 import {getFieldsForContentType} from "../src/contentTypeConfig";
 import {fetchSanity} from "../src/services/sanity";
 import {log} from './log.js';
-import {LocalStorageMock} from "./localStorageMock";
-
-const railContentModule = require('../src/services/railcontent.js')
+import {initializeTestService} from "./initializeTests";
 
 const {
     fetchSongById,
@@ -51,26 +48,6 @@ const {
 const {
     processMetadata,
 } = require('../src/contentMetaData.js');
-
-export function initializeTestService() {
-    const config = {
-        sanityConfig: {
-            token: process.env.SANITY_API_TOKEN,
-            projectId: process.env.SANITY_PROJECT_ID,
-            dataset: process.env.SANITY_DATASET,
-            useCachedAPI: process.env.SANITY_USE_CACHED_API === 'true' || true,
-            version: '2021-06-07',
-            debug: process.env.DEBUG === 'true' || false,
-            useDummyRailContentMethods: true,
-        },
-        localStorage: new LocalStorageMock()
-    };
-    initializeService(config);
-
-    let mock = jest.spyOn(railContentModule, 'fetchUserPermissionsData');
-    let testData = {"permissions": [78, 91, 92], "isAdmin": false};
-    mock.mockImplementation(() => testData);
-}
 
 describe('Sanity Queries', function () {
     beforeEach(() => {
